Changed up JSON and clean exiting.
Basically, now if the user tries to exit with ctrl+c it will stop exiting-->write to the JSON it is a clean exit-->Exit. This is to make sure the main file knows when to ask the user about next steps or to auto start.

If it exits unexpectedly the variable setting that it is "unexpected" will remain true, and make the program restart with clusters.